requirements.txt
torch
pyyaml
numpy
jsonschema
